added shouldSkipImagePath

+++ lib/ios.dart
@@ -52,6 +52,7 @@ final List<_IosLaunchImageTemplate> _iOSBrandingImagesDark =
 
 /// Create iOS splash screen
 void _createiOSSplash({
+  bool shouldSkipImagePath = true,
   required String? imagePath,
   required String? darkImagePath,
   String? brandingImagePath,
@@ -62,34 +63,35 @@ void _createiOSSplash({
   required String iosContentMode,
   String? iosBrandingContentMode,
   required bool fullscreen,
-  bool shouldCreateImagePath = false;
   required String? backgroundImage,
   required String? darkBackgroundImage,
 }) {
-  if (imagePath != null) {
-    _applyImageiOS(imagePath: imagePath, list: _iOSSplashImages);
-  } else if(shouldCreateImagePath) {
-    final splashImage = Image(width: 1, height: 1);
-    for (final template in _iOSSplashImages) {
-      final file =
-          File(_flavorHelper.iOSAssetsLaunchImageFolder + template.fileName);
-      file.createSync(recursive: true);
-      file.writeAsBytesSync(encodePng(splashImage));
+  if(!shouldSkipImagePath) {
+    if (imagePath != null) {
+      _applyImageiOS(imagePath: imagePath, list: _iOSSplashImages);
+    } else {
+      final splashImage = Image(width: 1, height: 1);
+      for (final template in _iOSSplashImages) {
+        final file =
+            File(_flavorHelper.iOSAssetsLaunchImageFolder + template.fileName);
+        file.createSync(recursive: true);
+        file.writeAsBytesSync(encodePng(splashImage));
+      }
     }
-  }
 
-  if (darkImagePath != null) {
-    _applyImageiOS(
-      imagePath: darkImagePath,
-      dark: true,
-      list: _iOSSplashImagesDark,
-    );
-  } else {
-    for (final template in _iOSSplashImagesDark) {
-      final file =
-          File(_flavorHelper.iOSAssetsLaunchImageFolder + template.fileName);
-      if (file.existsSync()) file.deleteSync();
-    }
+    if (darkImagePath != null) {
+      _applyImageiOS(
+        imagePath: darkImagePath,
+        dark: true,
+        list: _iOSSplashImagesDark,
+      );
+    } else {
+      for (final template in _iOSSplashImagesDark) {
+        final file =
+            File(_flavorHelper.iOSAssetsLaunchImageFolder + template.fileName);
+        if (file.existsSync()) file.deleteSync();
+      }
+    }  
   }
 
   if (brandingImagePath != null) {
